+++ src/aiida/orm/__init__.py
@@ -51,7 +51,6 @@ __all__ = (
     'Computer',
     'ComputerEntityLoader',
     'ContainerizedCode',
-    'ContractedProcessNode',
     'Data',
     'Dict',
     'Entity',