+++ src/aiida/orm/nodes/__init__.py
@@ -30,7 +30,6 @@ __all__ = (
     'CifData',
     'Code',
     'ContainerizedCode',
-    'ContractedProcessNode',
     'Data',
     'Dict',
     'EnumData',